requirements.txt
fastapi
uvicorn
FastAPI-SQLAlchemy
passlib
bcrypt
python-jose
python-multipart